refactor(sync): relocate docker sync-only files under sandbox-docker/src/sync/ (R2)
Co-location continued: the docker sync-only files now live under sync/. Pure
moves + import repointing, behaviour-preserving; public API (index) unchanged.

Moved into sandbox-docker/src/sync/: sync-transport, claude-credentials,
claude-hooks-filter, claude-pull, codex-config, dynamic-sync, host-export,
host-stage, in-box-git (docker-sync already there). home-ownership.ts stays
(container ownership-healing, not a sync op). Repointed create/lifecycle/
docker-provider/the per-tool files + index + the moved files' staying-file
imports + the test imports/vi.mock paths.

Gate: build + typecheck + lint green; docker 310 tests green.
