+++ packages/sandbox-docker/src/agents.ts
@@ -2,7 +2,7 @@ import { homedir } from 'node:os';
 import { join } from 'node:path';
 import { seedAgentsVolume } from '@agentbox/sandbox-core';
 import { ensureVolume, volumeExists } from './docker.js';
-import { createDockerSyncTransport } from './sync-transport.js';
+import { createDockerSyncTransport } from './sync/sync-transport.js';
 
 /**
  * The shared `~/.agents` directory — the cross-agent "Agent Skills" location
