+++ packages/sandbox-docker/src/sync/claude-hooks-filter.ts
