+++ packages/sandbox-docker/test/claude-hooks-filter.test.ts
@@ -5,7 +5,7 @@ import {
   filterHostHooks,
   isHostPathHookCommand,
   trustWorkspace,
-} from '../src/claude-hooks-filter.js';
+} from '../src/sync/claude-hooks-filter.js';
 
 describe('isHostPathHookCommand', () => {
   const home = '/Users/marco';
