+++ packages/sandbox-docker/test/scan-host-env-files.test.ts
@@ -2,7 +2,7 @@ import { mkdir, mkdtemp, rm, writeFile } from 'node:fs/promises';
 import { tmpdir } from 'node:os';
 import { join } from 'node:path';
 import { afterEach, beforeEach, describe, expect, it } from 'vitest';
-import { DEFAULT_ENV_PATTERNS, scanHostEnvFiles } from '../src/host-export.js';
+import { DEFAULT_ENV_PATTERNS, scanHostEnvFiles } from '../src/sync/host-export.js';
 
 describe('scanHostEnvFiles', () => {
   let dir: string;
